Revise and clean molecules and types
