+++ src/components/adhoc/molecules/StatsDisplay/StatsDisplay.tsx
@@ -2,7 +2,7 @@ import classes from './StatsDisplay.module.scss';
 
 import { StatBar } from '../../../../components/adhoc';
 
-import { IPokemon } from 'interfaces';
+import { IPokemon } from '../../../../interfaces';
 
 interface IProps {
   pokemon: IPokemon;
