+++ src/components/adhoc/molecules/TypesDisplay/TypesDisplay.tsx
@@ -2,19 +2,27 @@ import classes from './TypesDisplay.module.scss';
 
 import { TypeBadge } from '../../../../components/adhoc';
 
-import { POKEMON_TYPES } from 'enums';
+import { POKEMON_TYPES } from '../../../../enums';
 
 interface IProps {
   types: POKEMON_TYPES[];
 }
 
 const TypesDisplay = ({ types }: IProps) => {
-  return (
-    <div className={classes.mainContainer}>
-      <TypeBadge type={types[0]} />
-      {types[1] && <TypeBadge type={types[1]} />}
-    </div>
-  );
+  const renderTypeBadges = () => {
+    if (types[1]) {
+      return (
+        <>
+          <TypeBadge type={types[0]} />
+          <TypeBadge type={types[1]} />
+        </>
+      );
+    } else {
+      return <TypeBadge type={types[0]} />;
+    }
+  };
+
+  return <div className={classes.mainContainer}>{renderTypeBadges()}</div>;
 };
 
 export default TypesDisplay;
