+++ src/interfaces/carousel.ts
@@ -1,0 +1,11 @@
+export interface IPageClassNames {
+  pageA: string;
+  pageB: string;
+  pageC: string;
+}
+
+export interface IPageIdLists {
+  pageA: number[];
+  pageB: number[];
+  pageC: number[];
+}
