+++ src/interfaces/generic.ts
@@ -1,0 +1,3 @@
+export interface IClassNames {
+  readonly [key: string]: string;
+}
